# Migrate kl_div_backward from CUDA_tensor_apply3 to TensorIterator