Stack from [ghstack](https://github.com/ezyang/ghstack):
* #34026 Completely kill CUDA_tensor_apply3
* #34025 Migrate lerp from CUDA_tensor_apply3 to TensorIterator
* #34024 Migrate dropout inference from CUDA_tensor_apply3 to TensorIterator
* #34023 Migrate bce loss from CUDA_tensor_apply3 to TensorIterator
* **#34022 Migrate kl_div_backward from CUDA_tensor_apply3 to TensorIterator**
* #34021 Migrate dirichlet from CUDA_tensor_apply3 to TensorIterator
* #34020 Migrate gamma grad from CUDA_tensor_apply3 to TensorIterator

Differential Revision: [D20196080](https://our.internmc.facebook.com/intern/diff/D20196080)