@zasdfgbnm 2020-03-02
@VitalyFedyunin On small tensors, the performance is improved. On large tensors, the performance are the same.

```python
import torch
print(torch.__version__)

# warm up
for i in range(1000):
    torch.arange(100000, device='cuda')
    
def benchmark(m, size):
    s1 = 2 ** (size // 2)
    s2 = 2 ** (size - size // 2)
    a = torch.zeros((s1, s2), device='cuda')
    b = torch.ones_like(a, requires_grad=True)
    loss = m(a, b)
    torch.cuda.synchronize()
    %timeit loss.backward(retain_graph=True); torch.cuda.synchronize()

print("reduction = mean:")
m = torch.nn.KLDivLoss().cuda()
for i in range(15, 28):
    benchmark(m, i)

print()
print("reduction = sum:")
m2 = torch.nn.KLDivLoss(reduction="sum").cuda()
for i in range(15, 28):
    benchmark(m2, i)
```

Before:
```
1.5.0a0+9956a23
reduction = mean:
179 µs ± 2.05 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
177 µs ± 2.99 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
173 µs ± 568 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
215 µs ± 4.64 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
376 µs ± 3.43 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
678 µs ± 478 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
1.29 ms ± 650 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
2.55 ms ± 15.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
5.01 ms ± 16.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
9.96 ms ± 733 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
19.8 ms ± 990 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
39.4 ms ± 56.9 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
78.7 ms ± 43.1 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

reduction = sum:
164 µs ± 1.01 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
159 µs ± 5.19 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
154 µs ± 150 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
209 µs ± 11.2 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
332 µs ± 2.29 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
592 µs ± 1.25 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
1.13 ms ± 337 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
2.21 ms ± 5.12 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
4.37 ms ± 13.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
8.72 ms ± 3.71 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
17.3 ms ± 5.41 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
34.5 ms ± 49.9 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
68.9 ms ± 52.8 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

After:
```
1.5.0a0+fdac34c
reduction = mean:
113 µs ± 32.1 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
159 µs ± 2.79 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
162 µs ± 34.2 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
215 µs ± 619 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
372 µs ± 1.04 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
664 µs ± 438 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
1.28 ms ± 599 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
2.51 ms ± 4.95 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
4.99 ms ± 5.42 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
9.95 ms ± 554 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
19.8 ms ± 866 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
39.4 ms ± 56.1 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
78.7 ms ± 46.4 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

reduction = sum:
142 µs ± 336 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
142 µs ± 222 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
143 µs ± 1.74 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
191 µs ± 1.31 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
328 µs ± 575 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
585 µs ± 523 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
1.12 ms ± 384 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
2.2 ms ± 5.02 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
4.38 ms ± 5.69 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
8.72 ms ± 808 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
17.3 ms ± 4.93 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
34.5 ms ± 49.3 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
68.9 ms ± 51.8 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```